+++ fabric/src/main/java/tech/alexnijjar/extractinator/common/config/fabric/ModMenuConfig.java
@@ -1,7 +1,7 @@
 package tech.alexnijjar.extractinator.common.config.fabric;
 
+import com.teamresourceful.resourcefulconfig.api.types.ResourcefulConfig;
 import com.teamresourceful.resourcefulconfig.client.ConfigScreen;
-import com.teamresourceful.resourcefulconfig.common.config.ResourcefulConfig;
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
 import net.fabricmc.api.EnvType;